cleanup classes (use lombok), add import export of hierarchy: need to be corrected

+++ src/main/java/com/aquila/ibm/mq/gui/model/QueueManagerConfig.java
@@ -37,6 +37,6 @@ public class QueueManagerConfig {
     }
 
     public String getLabel() {
-        return String.format("%s(%d)", host, port);
+        return String.format("%s(%d)", host, (Integer)port);
     }
 }
